chore: pr comments
Updates the ARC56 contract data structures to improve clarity and consistency.

Renames `StructFieldType::Simple` to `StructFieldType::Value` to better reflect its purpose.
Renames `CallEnum` to `CallOnApplicationComplete` and `CreateEnum` to `CreateOnApplicationComplete` to explicitly express the purpose of the enums.
Renames `Global` and `Local` to `StateSchema` and uses the same `StateSchema` struct for both global and local state definitions.